Fix displaying transactions from Quipuswap
