+++ src/app/templates/OperationHistory.tsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 import classNames from "clsx";
 import BigNumber from "bignumber.js";
-import formatDistanceToNow from "date-fns/formatDistanceToNow";
 import { useRetryableSWR } from "lib/swr";
 import {
   TZSTATS_CHAINS,
@@ -10,7 +9,7 @@ import {
   TZStatsOperation,
 } from "lib/tzstats";
 import { loadChainId } from "lib/thanos/helpers";
-import { T, TProps } from "lib/i18n/react";
+import { T } from "lib/i18n/react";
 import {
   ThanosAssetType,
   ThanosAsset,
@@ -20,7 +19,6 @@ import {
   useOnStorageChanged,
   mutezToTz,
   isKnownChainId,
-  useAssets,
   ThanosToken,
   useChainId,
 } from "lib/thanos/front";
@@ -31,36 +29,14 @@ import {
   BcdTokenTransfer,
   getTokenTransfers,
 } from "lib/better-call-dev";
-import InUSD from "app/templates/InUSD";
-import HashChip from "app/templates/HashChip";
-import Identicon from "app/atoms/Identicon";
-import OpenInExplorerChip from "app/atoms/OpenInExplorerChip";
-import Money from "app/atoms/Money";
 import { ReactComponent as LayersIcon } from "app/icons/layers.svg";
-import { ReactComponent as ClipboardIcon } from "app/icons/clipboard.svg";
+import Operation, {
+  OperationPreview,
+  InternalTransfer,
+} from "app/templates/Operation";
 
 const PNDOP_EXPIRE_DELAY = 1000 * 60 * 60 * 24;
 const OPERATIONS_LIMIT = 30;
-
-interface InternalTransfer {
-  volume: BigNumber;
-  tokenId?: number;
-  sender: string;
-  receiver: string;
-  tokenAddress?: string;
-}
-
-interface OperationPreview {
-  entrypoint?: string;
-  rawReceiver?: string;
-  delegate?: string;
-  hash: string;
-  type: string;
-  status: string;
-  time: string;
-  internalTransfers: InternalTransfer[];
-  volume: number;
-}
 
 interface OperationHistoryProps {
   accountPkh: string;
@@ -203,11 +179,11 @@ const AllOperationsList: React.FC<AllOperationsListProps> = ({
 
       return [
         ...Object.keys(tzStatsOps)
-          .reduce<OperationPreview[]>((mergedOps, opHash) => {
+          .reduce<OperationPreview[]>((prevMergedOps, opHash) => {
             const opsChunk = tzStatsOps[opHash];
             const bcdOpsChunk = bcdOps[opHash] || [];
-            return opsChunk.reduce<OperationPreview[]>(
-              (mergedOpsFromChunk, op) => {
+            const mergedOps = opsChunk.reduce<OperationPreview[]>(
+              (mergedOpsFromChunk, op, index) => {
                 const transfersFromParams =
                   op.type === "transaction" &&
                   (op.parameters as any)?.entrypoint === "transfer"
@@ -215,7 +191,7 @@ const AllOperationsList: React.FC<AllOperationsListProps> = ({
                     : null;
                 const transfersFromVolumeProp =
                   op.type === "transaction" &&
-                  (!op.parameters || bcdOps[op.hash])
+                  (!op.parameters || (bcdOps[op.hash] && op.volume > 0))
                     ? [
                         {
                           volume: new BigNumber(op.volume),
@@ -224,43 +200,61 @@ const AllOperationsList: React.FC<AllOperationsListProps> = ({
                         },
                       ]
                     : [];
-                return [
-                  ...mergedOpsFromChunk,
-                  {
-                    delegate:
-                      op.type === "delegation" ? op.delegate : undefined,
-                    entrypoint: (op.parameters as any)?.entrypoint,
-                    internalTransfers: [
-                      ...(transfersFromParams || transfersFromVolumeProp).map(
-                        (transfer) => ({
-                          ...transfer,
-                          tokenAddress: transfersFromParams
-                            ? op.receiver
-                            : undefined,
-                        })
-                      ),
-                      ...(transfersFromParams?.length === bcdOpsChunk.length
-                        ? []
-                        : bcdOpsChunk
-                      ).map((bcdOp) => ({
-                        volume: new BigNumber(bcdOp.amount),
-                        tokenId: bcdOp.token_id,
-                        sender: bcdOp.from,
-                        receiver: bcdOp.to,
-                        tokenAddress: bcdOp.contract,
-                      })),
-                    ],
-                    hash: op.hash,
-                    status: op.status,
-                    time: op.time,
-                    type: op.type,
-                    volume: op.volume,
-                    rawReceiver: op.receiver,
-                  },
+                const internalTransfers = [
+                  ...(transfersFromParams || transfersFromVolumeProp).map(
+                    (transfer) => ({
+                      ...transfer,
+                      tokenAddress: transfersFromParams
+                        ? op.receiver
+                        : undefined,
+                    })
+                  ),
+                  ...((transfersFromParams?.length || 0) >=
+                    bcdOpsChunk.length || index > 0
+                    ? []
+                    : bcdOpsChunk
+                  ).map((bcdOp) => ({
+                    volume: new BigNumber(bcdOp.amount),
+                    tokenId: bcdOp.token_id,
+                    sender: bcdOp.from,
+                    receiver: bcdOp.to,
+                    tokenAddress: bcdOp.contract,
+                  })),
                 ];
+                const transactionIndex = mergedOpsFromChunk.findIndex(
+                  ({ type }) => type === "transaction"
+                );
+                if (op.type !== "transaction" || transactionIndex === -1) {
+                  return [
+                    ...mergedOpsFromChunk,
+                    {
+                      counter: op.counter,
+                      delegate:
+                        op.type === "delegation" ? op.delegate : undefined,
+                      entrypoint: (op.parameters as any)?.entrypoint,
+                      internalTransfers,
+                      hash: op.hash,
+                      status: op.status,
+                      time: op.time,
+                      type: op.type,
+                      volume: op.volume,
+                      rawReceiver: op.receiver,
+                    },
+                  ];
+                }
+                const oldTransaction = mergedOpsFromChunk[transactionIndex];
+                mergedOpsFromChunk[transactionIndex] = {
+                  ...mergedOpsFromChunk[transactionIndex],
+                  internalTransfers: [
+                    ...oldTransaction.internalTransfers,
+                    ...internalTransfers,
+                  ],
+                };
+                return mergedOpsFromChunk;
               },
-              mergedOps
+              []
             );
+            return [...prevMergedOps, ...mergedOps];
           }, [])
           .filter(({ volume, type, entrypoint }) => {
             if (!xtzOnly) {
@@ -441,7 +435,7 @@ const GenericOperationsList: React.FC<GenericOperationsListProps> = ({
   const pendingOperations = React.useMemo<OperationPreview[]>(
     () =>
       pndOps
-        .map((op) => {
+        .map((op, index) => {
           const parameters = (op as any).parameters;
           let internalTransfers: InternalTransfer[] = [];
           let tokenAddress: string | undefined = undefined;
@@ -464,6 +458,8 @@ const GenericOperationsList: React.FC<GenericOperationsListProps> = ({
 
           return {
             ...op,
+            // @ts-ignore
+            counter: Number(op.counter || index),
             entrypoint: parameters?.entrypoint,
             hash: op.hash,
             type: op.kind,
@@ -568,398 +564,12 @@ const GenericOperationsList: React.FC<GenericOperationsListProps> = ({
   );
 };
 
-type OperationProps = OperationPreview & {
-  accountPkh: string;
-  withExplorer: boolean;
-  explorerBaseUrl: string | null;
-};
-
-const Operation = React.memo<OperationProps>(
-  ({
-    accountPkh,
-    delegate,
-    entrypoint,
-    withExplorer,
-    explorerBaseUrl,
-    hash,
-    rawReceiver,
-    type,
-    status,
-    time,
-    internalTransfers,
-    volume,
-  }) => {
-    const imReceiver = internalTransfers.some(
-      ({ receiver }) => receiver === accountPkh
-    );
-    const pending = withExplorer && status === "pending";
-    const failed = ["failed", "backtracked", "skipped"].includes(status);
-    const volumeExists = volume > 0;
-    const hasTokenTransfers = internalTransfers.some(
-      ({ tokenAddress }) => !!tokenAddress
-    );
-    const sender = internalTransfers[0]?.sender;
-    const hasReceival = internalTransfers.some(
-      ({ receiver }) => receiver === accountPkh
-    );
-    const hasSending = internalTransfers.some(
-      ({ sender }) => sender === accountPkh
-    );
-    const isTransfer =
-      (hasTokenTransfers || (volumeExists && type === "transaction")) &&
-      (!entrypoint || entrypoint === "transfer") &&
-      !(internalTransfers.length > 1 && hasSending && hasReceival);
-    const isSendingTransfer = isTransfer && !imReceiver;
-    const isReceivingTransfer = isTransfer && imReceiver;
-    const moreExactType = React.useMemo(() => {
-      const rawReceiverIsContract =
-        !!rawReceiver && rawReceiver.startsWith("KT");
-      const isMultipleTransfersInteraction =
-        internalTransfers.length > 1 &&
-        internalTransfers.some(({ sender }) => sender.startsWith("KT")) &&
-        internalTransfers.some(({ receiver }) => receiver.startsWith("KT"));
-      switch (true) {
-        case isTransfer:
-          return "transfer";
-        case type === "delegation":
-          return delegate ? "delegation" : "undelegation";
-        case type === "transaction" &&
-          (rawReceiverIsContract || isMultipleTransfersInteraction):
-          return "interaction";
-        default:
-          return type;
-      }
-    }, [isTransfer, rawReceiver, type, delegate, internalTransfers]);
-
-    const receivers = React.useMemo(() => {
-      const uniqueReceivers = new Set(
-        internalTransfers.map((transfer) => transfer.receiver)
-      );
-      return [...uniqueReceivers];
-    }, [internalTransfers]);
-
-    const { iconHash, iconType } = React.useMemo<{
-      iconHash: string;
-      iconType: "bottts" | "jdenticon";
-    }>(() => {
-      switch (true) {
-        case isSendingTransfer:
-          return { iconHash: receivers[0], iconType: "bottts" };
-        case isReceivingTransfer:
-          return { iconHash: sender, iconType: "bottts" };
-        case type === "delegation" && !!delegate:
-          return { iconHash: delegate!, iconType: "bottts" };
-        case moreExactType === "interaction":
-          return { iconHash: rawReceiver!, iconType: "jdenticon" };
-        default:
-          return { iconHash: hash, iconType: "jdenticon" };
-      }
-    }, [
-      delegate,
-      hash,
-      type,
-      moreExactType,
-      isReceivingTransfer,
-      receivers,
-      isSendingTransfer,
-      rawReceiver,
-      sender,
-    ]);
-
-    return (
-      <div className={classNames("my-3", "flex items-stretch")}>
-        <div className="mr-2">
-          <Identicon
-            hash={iconHash}
-            type={iconType}
-            size={50}
-            className="shadow-xs"
-          />
-        </div>
-
-        <div className="flex-1">
-          <div className="flex items-center">
-            <HashChip
-              hash={hash}
-              firstCharsCount={10}
-              lastCharsCount={7}
-              small
-              className="mr-2"
-            />
-
-            {explorerBaseUrl && (
-              <OpenInExplorerChip
-                baseUrl={explorerBaseUrl}
-                opHash={hash}
-                className="mr-2"
-              />
-            )}
-
-            <div className={classNames("flex-1", "h-px", "bg-gray-200")} />
-          </div>
-
-          <div className="flex items-stretch">
-            <div className="flex flex-col">
-              <div className="flex items-center mt-1 text-xs text-blue-600 opacity-75">
-                {formatOperationType(moreExactType, imReceiver)}
-              </div>
-              {isReceivingTransfer && (
-                <OperationArgumentDisplay
-                  i18nKey="transferFromSmb"
-                  arg={[sender]}
-                />
-              )}
-              {isSendingTransfer && (
-                <OperationArgumentDisplay
-                  i18nKey="transferToSmb"
-                  arg={receivers}
-                />
-              )}
-              {moreExactType === "interaction" && (
-                <OperationArgumentDisplay
-                  i18nKey="interactionWithContract"
-                  arg={[rawReceiver!]}
-                />
-              )}
-              {moreExactType === "delegation" && (
-                <OperationArgumentDisplay
-                  i18nKey="delegationToSmb"
-                  arg={[delegate!]}
-                />
-              )}
-
-              {(() => {
-                const timeNode = (
-                  <Time
-                    children={() => (
-                      <span className="text-xs font-light text-gray-500">
-                        {formatDistanceToNow(new Date(time), {
-                          includeSeconds: true,
-                          addSuffix: true,
-                          // @TODO Add dateFnsLocale
-                          // locale: dateFnsLocale,
-                        })}
-                      </span>
-                    )}
-                  />
-                );
-
-                switch (true) {
-                  case failed:
-                    return (
-                      <div className="flex items-center">
-                        <T id={status}>
-                          {(message) => (
-                            <span className="mr-1 text-xs font-light text-red-700">
-                              {message}
-                            </span>
-                          )}
-                        </T>
-
-                        {timeNode}
-                      </div>
-                    );
-
-                  case pending:
-                    return (
-                      <T id="pending">
-                        {(message) => (
-                          <span className="text-xs font-light text-yellow-600">
-                            {message}
-                          </span>
-                        )}
-                      </T>
-                    );
-
-                  default:
-                    return timeNode;
-                }
-              })()}
-            </div>
-
-            <div className="flex-1" />
-
-            {!failed && (
-              <div className="flex flex-col items-end flex-shrink-0">
-                {internalTransfers.map((transfer, index) => (
-                  <OperationVolumeDisplay
-                    accountPkh={accountPkh}
-                    key={index}
-                    tokenAddress={transfer.tokenAddress}
-                    transfer={transfer}
-                    pending={pending}
-                    volume={0}
-                  />
-                ))}
-                {internalTransfers.length === 0 &&
-                  (volume || undefined) &&
-                  (isTransfer || type === "delegation") && (
-                    <OperationVolumeDisplay
-                      accountPkh={accountPkh}
-                      pending={pending}
-                      volume={volume}
-                    />
-                  )}
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-    );
-  }
-);
-
-type OperationArgumentDisplayProps = {
-  i18nKey: TProps["id"];
-  arg: string[];
-};
-
-const OperationArgumentDisplay = React.memo<OperationArgumentDisplayProps>(
-  ({ i18nKey, arg }) => (
-    <span className="font-light text-gray-500 text-xs">
-      <T
-        id={i18nKey}
-        substitutions={arg.map((value, index) => (
-          <>
-            <HashChip
-              className="text-blue-600 opacity-75"
-              key={index}
-              hash={value}
-              type="link"
-            />
-            {index === arg.length - 1 ? null : ", "}
-          </>
-        ))}
-      />
-    </span>
-  )
-);
-
-type OperationVolumeDisplayProps = {
-  accountPkh: string;
-  tokenAddress?: string;
-  transfer?: InternalTransfer;
-  pending: boolean;
-  volume: number;
-};
-
-const OperationVolumeDisplay: React.FC<OperationVolumeDisplayProps> = (
-  props
-) => {
-  const { accountPkh, tokenAddress, pending, transfer, volume } = props;
-
-  const { allAssets } = useAssets();
-  const token = React.useMemo(() => {
-    if (!transfer || !tokenAddress) {
-      return undefined;
-    }
-    return allAssets.find((a) => {
-      return (
-        a.type !== ThanosAssetType.XTZ &&
-        a.address === tokenAddress &&
-        (a.type !== ThanosAssetType.FA2 || a.id === transfer.tokenId)
-      );
-    });
-  }, [allAssets, transfer, tokenAddress]);
-
-  const finalVolume = transfer
-    ? transfer.volume.div(10 ** (token?.decimals || 0))
-    : volume;
-  const type = React.useMemo(() => {
-    if (transfer) {
-      return transfer.receiver === accountPkh ? "receive" : "send";
-    }
-    return "other";
-  }, [accountPkh, transfer]);
-  const isTransaction = type === "receive" || type === "send";
-
-  return (
-    <>
-      <div
-        className={classNames(
-          "text-sm",
-          (() => {
-            switch (true) {
-              case pending:
-                return "text-yellow-600";
-
-              case isTransaction:
-                return type === "receive" ? "text-green-500" : "text-red-700";
-
-              default:
-                return "text-gray-800";
-            }
-          })()
-        )}
-      >
-        {isTransaction && (type === "receive" ? "+" : "-")}
-        {BigNumber.isBigNumber(finalVolume) && finalVolume.isNaN() ? (
-          "?"
-        ) : (
-          <Money>{finalVolume}</Money>
-        )}{" "}
-        {tokenAddress ? token?.symbol || "???" : "ꜩ"}
-      </div>
-
-      {(!tokenAddress || token) && (
-        <InUSD volume={finalVolume} asset={token || XTZ_ASSET}>
-          {(usdVolume) => (
-            <div className="text-xs text-gray-500">
-              <span className="mr-px">$</span>
-              {usdVolume}
-            </div>
-          )}
-        </InUSD>
-      )}
-    </>
-  );
-};
-
-type TimeProps = {
-  children: () => React.ReactElement;
-};
-
-const Time: React.FC<TimeProps> = ({ children }) => {
-  const [value, setValue] = React.useState(children);
-
-  React.useEffect(() => {
-    const interval = setInterval(() => {
-      setValue(children());
-    }, 5_000);
-
-    return () => {
-      clearInterval(interval);
-    };
-  }, [setValue, children]);
-
-  return value;
-};
-
-function formatOperationType(type: string, imReciever: boolean) {
-  if (type === "transaction" || type === "transfer") {
-    type = `${imReciever ? "↓" : "↑"}_${type}`;
-  }
-
-  const operationTypeText = type
-    .split("_")
-    .map((w) => `${w.charAt(0).toUpperCase()}${w.substring(1)}`)
-    .join(" ");
-
-  return (
-    <>
-      {type === "interaction" && (
-        <ClipboardIcon className="mr-1 h-3 w-auto stroke-current inline align-text-top" />
-      )}
-      {operationTypeText}
-    </>
-  );
-}
-
 function opKey(op: OperationPreview) {
-  return `${op.hash}_${op.type}`;
+  return `${op.hash}_${op.type}_${op.counter}`;
 }
 
 function tryGetTransfers(parameters: any): InternalTransfer[] | null {
+  console.log(parameters);
   try {
     if (parameters.value instanceof Array) {
       const parsedTransfers: InternalTransfer[] = [];
